[Tab Model] add tabGroupExists method
The tabGroupExistsForRootId method is deprecated and tabGroupExists is
introduced. The existing method is strange in that it reported any
root ID with a TabGroup object as having a group when in reality groups
with a single tab may not be "real" tab groups. So it was more to check
if TabGroupModelFilter was tracking the root ID at all. The new method
will instead check for a "valid" tab group which makes more sense.

In follow-up CLs if any callers relied on the prior behavior then
alternative mechanisms might be required, but that would imply callers
relied on an internal implementation detail of TabGroupModelFilterImpl
so hopefully that is not the case.

Bug: 399354986
Change-Id: I2ad0f6fee1453650e7d52ff4b652f08fbde3a395
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6309795
Auto-Submit: Calder Kitagawa <ckitagawa@chromium.org>
Commit-Queue: Sky Malice <skym@chromium.org>
Reviewed-by: Sky Malice <skym@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1425802}