Uprev ChromeScreenAI CIPD to 127.19
This CL updates the CIPD for downloading the library that is used for
ChromeScreenAI browser tests.

If browser tests fail, revert the rollout for the new binary in the
component updater as well. You will need to rollout a newer (or older)
release to update the clients that are affected by the faulty rollout.

Bug: 378472917

Change-Id: I1e277610f8ab708fad37dcdaf620445b6fd0d935
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6304019
Commit-Queue: Ramin Halavati <rhalavati@chromium.org>
Reviewed-by: Ramin Halavati <rhalavati@chromium.org>
Reviewed-by: Mark Schillaci <mschillaci@google.com>
Reviewed-by: Kyungjun Lee <kyungjunlee@google.com>
Cr-Commit-Position: refs/heads/main@{#1425818}