[MIM2]Do not show management confirmation after it has been accepted
This is a follow-up to crrev.com/c/6243097, which adds better support
for multi profiles. If a managed profile is fully initialized, it means
that the user has already seen the seen the management confirmation and
accepted it, so they should never see it again.

Bug: 395086006
Change-Id: I68dabe4422be06080f5c6295812405dbe55f8290
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6297809
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Yann Dago <ydago@chromium.org>
Auto-Submit: Yann Dago <ydago@chromium.org>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1425842}