[Metadata][Metrics] Remove negative bias from image fetcher metrics
On Android devices, Autofill infra triggers fetching of all existing
credit card art images on browser startup. These images are cached in
memory. When a UI client attempts to get these images, the infra retries
fetching if it's not found in cache. This inflates the image fetching
failure rate because the image fetch attempt is made every time the
image needs to be shown.

Do not retry fetching if the card art image was not found in cache.
This change is already implemented for Desktop clients.

Bug: 40276036
Change-Id: I9ea147a1b050468ec38eac8b721e0add38515a67
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6244007
Reviewed-by: Siddharth Shah <siashah@chromium.org>
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Commit-Queue: Vishwas Uppoor <vishwasuppoor@google.com>
Cr-Commit-Position: refs/heads/main@{#1425845}