Add Probabilistic Reveal Token registry
The PRT registry holds the list of domains that should receive PRTs in
their request headers. The registry is populated by the component
updater and is owned by the Network Service.

The IpProtectionCore class exposes a new method for determining whether
a given request should include the PRT header by checking if the
requested domain is in the registry.

Bug: 396401608
Change-Id: I194c361a57695a0652c40d566a3dd2e131fc0337
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6278626
Reviewed-by: Dominic Farolino <dom@chromium.org>
Reviewed-by: Dustin Mitchell <djmitche@chromium.org>
Commit-Queue: Ryan Kalla <ryankalla@google.com>
Reviewed-by: Sorin Jianu <sorin@chromium.org>
Reviewed-by: Aykut Bulut <aykutb@google.com>
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1425863}