Add deprecation notices to test reproduction tools
With the development and use of the UTR as far long as it is, these
old tools have mostly been obviated. So this adds a deprecation
notice to run-swarmed.py as well as "mb.py run". This notice makes the
scripts fail, unless --force is passed in. Plan is to let this sit
for a while, then eventually delete the scripts.

Bug: 386167803
Change-Id: I40a663a1ce4ea6d64d75a8007d14024bf9319d53
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6127280
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@google.com>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Cr-Commit-Position: refs/heads/main@{#1425877}