Add XRWebGLSwapChain backed by a Drawing Buffer
Creates a new type of swap chain based on an XRWebGLDrawingBuffer, which
makes it much closer to the XRWebGLLayer. Only used when
DrawingIntoSharedBuffer is false. This allows us to temporarily avoid
the synchronization issues we're seeing on Windows with the SharedImage
code path, allowing us to ship faster.

Bug: 40700985
Change-Id: I2069ba02cbafe7bce750e192bbb34dcbd6bdd203
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6306886
Reviewed-by: Alexander Cooper <alcooper@chromium.org>
Commit-Queue: Brandon Jones <bajones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1425898}