Pad snackbars in SearchActivity
Currently, SearchActivity initializes its SnackbarManager with android.R.id.content as the parent view for snackbars. This causes the snackbars to draw into window insets when drawing edge-to-edge - if edge-to-edge is enabled, the EdgeToEdgeBaseLayout should be used instead, as it automatically pads the content to account for window insets / system bars.

Before:
https://screenshot.googleplex.com/4LvHFvCKX6NBFsc.png

After:
https://screenshot.googleplex.com/6RdVjnViXYzxEoE.png

Bug: 398937046
Change-Id: I797803ea1571d7ab0ec22a6fab86994350811791
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6299813
Commit-Queue: Charles Hager <clhager@google.com>
Reviewed-by: Tomasz Wiszkowski <ender@google.com>
Cr-Commit-Position: refs/heads/main@{#1425904}