Dynamically add CSS gradient
This CL moves some .css to a file that is pulled from a resource at
runtime. Unlike other .css which gets built into css.js files, this
is pulled in via a link tag, which is a bit clunky but does permit
this flexibility.

Bug: 398322302
Change-Id: I1f8b63121c0d32abae61a210888d2f13e8158a6f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6304104
Reviewed-by: Erik Chen <erikchen@chromium.org>
Commit-Queue: Ian Vollick <vollick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1425953}