media/hls: migrate crypto code to new //crypto APIs
This change migrates the //media HLS implementation to the new //crypto
APIs, dropping deps on crypto/encryptor and crypto/symmetric_key and
adding deps on crypto/aes_cbc and crypto/random. This allows removing
some heap allocations & copies and adding some memory safety, as well as
dropping deps on the deprecated APIs.

In places the new versions are more verbose because the rest of the CDM
code uses std::string for plaintexts and ciphertexts, whereas the new
//crypto APIs tend to use spans and std::vectors. That leads to some
(cheap) conversions at call sites to/from base::spans which clutter the
code a little. It would be nice if the CDM code used
std::vector<uint8_t> for this data instead but that doesn't seem worth
doing presently.

Change-Id: I1a539c7f26119235c00c0f272551da441bd088e4
Bug: 372283556
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6191414
Commit-Queue: Elly FJ <ellyjones@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: David Benjamin <davidben@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1426056}