[lensoverlay] Show toast when text is not found at link.
This adds a function to the LensSidePanelPage to show a toast with a
message provided in the side panel. This is currently used to show a
toast when the text finder could not find the text in the tab but it is
the same URL as what was clicked in the side panel.

Change-Id: I587c0b839a6d0c7195081b59750faedb73ae97ac
Bug: 399452472
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6307127
Reviewed-by: Ali Stanfield <stanfield@google.com>
Commit-Queue: Juan Mojica <juanmojica@google.com>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1426062}