Clean up RPH usage in RenderProcessHostTaskProvider::CreateTask()
Currently, the CreateTask() callers have a RenderProcessHost pointer,
but they pass in an ID. Then CreateTask() takes the ID and converts it
back to a pointer. Remove this extra work and just pass in the pointer.

Modernize comments to use backticks along the way.

Change-Id: I88b22b7afffed8e898f6aca5469646f639da625a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6310441
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1426106}