[Tab Group Sync] Mark remote originating group as hidden
This CL fixes the following case when a shared tab group is applied
first, and then there is a remote creation of the originating saved tab
group. For example, the same order might happen on sign-in.

This CL also fixes handling empty groups which was always the case for
remotely created groups (they are always empty at the first step). In
the service, such remote groups might bypass some additional conditions
(e.g. whether the group is hidden or transitioned).

This CL also adds an integration test to cover this order of remote
creations of groups.

Bug: 399461032
Change-Id: I673d192f811a8240ba37e9c48de90c2315890291
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6304691
Reviewed-by: Shakti Sahu <shaktisahu@chromium.org>
Commit-Queue: Rushan Suleymanov <rushans@google.com>
Cr-Commit-Position: refs/heads/main@{#1426193}