[cc/scheduler] Align main frame throttling on Impl frame timing
Prior to this CL, main frame throttling relied on TimeTicks::Now() when
the main frame is about to be sent. This is thus subject to jitter if
and when there is descheduling, or long processing delays. From a code
cleanliness standpoint, it also makes SchedulerStateMachine depend on
hidden state, i.e. current time.

This CL makes all throttling decisions based on the timestamp contained
in BeginFrameArgs. It eliminates the source of jitter, and makes all
computations deterministic.

Bug: 379678455
Change-Id: I1127e72faa50c1dd9b13d815ec18862805da66b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6304752
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Commit-Queue: Benoit Lize <lizeb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1426202}