[functions] Don't std::move argument/type maps around
Instead of FunctionContext holding `arguments` and `local_types`
by value, just refer to objects to the stack with a const reference.
This makes it more clear what can and cannot be modified
on a FunctionContext, and removes the need to awkwardly std::move
things back and forth.

There should be no behavior change.

Bug: 397459622, 325504770
Change-Id: I86e2aca0f94adf7a29b77ff96d8f61e1892cdbd0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6300495
Reviewed-by: Steinar H Gunderson <sesse@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1426216}