[UPMLoginDbSunset] Introduce CSV download dialog
This CL introduces all the scaffolding for the CSV download component
and wires first part of the flow.

The flow is as follows:
1. A modal dialog is shown with explanations on why GPM is unavailable
and a way to download unmigrated passwords.
2. If the user clicks "Export passwords", reauthentication is
triggered.
   2a. If reauth isn't available a toast is shown telling the user to set up screen lock.
   2b. If reauth fails, the flow ends.
   2c. If reauth succeeds, the flow proceeds with step 3.
3. An activity is launched that allows the user to choose the location
and name for the file. It returns a URI to the document to write in.

The flow will continue with file write in the next CL.



Change-Id: Ie23177df69d6c7e41e72b98112301c027a509276
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6304709
Reviewed-by: Friedrich Hauser <friedrichh@chromium.org>
Commit-Queue: Ioana Pandele <ioanap@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1426221}