[cc] Trigger ImplFrame deadline immediately when main is throttled
When a BeginMainFrame() is not sent, do not wait for the Impl frame
deadline. Otherwise, we falsely believe that we dropped a frame when the
deadline comes without a commit.

One example where this always happens is when a page has rAF animations,
but is otherwise static, with main running at 60 and impl at 120. In
this case, every other (compositor) frame has no update, and does not
send BeginMainFrame(). Without this CL, we still post a deadline for all
these frames, which is wasteful, and can inflate missed frame metrics
too.

This commit does not go all the way: removing the "needs_redraw_"
condition was attempted, and reverted in the past. See the TODO in the
code for details.

Bug: 379678455
Change-Id: I3210cd815f0bbef5b2129f54b89e40c2187dd6f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6309083
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Commit-Queue: Benoit Lize <lizeb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1426226}