Remove line_bounds from ImeCompositionRangeChanged
The line bounds are only used on Android and we no longer use this code
path on Android (as of https://crrev.com/c/5132102). We will need to
make the same change for composition character bounds. The reason these
are being removed separately is mostly for caution.

Bug: 40940885
Change-Id: I6ce8367a140f79600adda193af52a45e2f4d8468
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6207093
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Alex Mitra <alexmitra@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Antonio Sartori <antoniosartori@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1426268}