Fix empty item DCHECK failure in LineInfo::ComputeTrailingSpaceWidth
While working on crrev.com/c/6298873, I noticed a similar DCHECK in
`LineInfo::ComputeTrailingSpaceWidth` which could lead to failures.
As in crrev.com/c/6298873, this DCHECK was testing that an item result
could not have a zero end offset, which can happen with an empty
control item (such as a U+000D carriage return in the DOM text) at the
start of the IFC. However, reproducing this failure was not as easy.

`LineInfo::ComputeTrailingSpaceWidth` computes the hanging width at
the end of the line (see https://drafts.csswg.org/css-text-4/#hang).
This method returns early (without reaching the DCHECK) if the line
doesn't have any preserved trailing spaces, so those are necessary for
the failure; however, most preserved trailing spaces (such as the ones
caused by `white-space: pre`) don't hang, which breaks the loop before
reaching the carriage return item. However, a case where they can hang
is if they have `white-space: pre-wrap` and the space overflows the
line. And this does in fact cause a DCHECK failure.

This patch fixes this by moving the DCHECK right before the code that
iterates through the characters in the item result, which needs that
assertion to hold. Since the function returns before this step if the
item result has zero length, this fixes that DCHECK failure.

Change-Id: I308cd1fe4b6ac7b2c0823eab976fe4f896a3c576
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6300503
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: Andreu Botella <abotella@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1426274}