[iOS] Create Scheduler Class
This CL implements the Scheduler class. The scheduler is responsible for
managing and updating the ScheduledFiles contained within the
ScheduledFileQueue. In a follow-up CL, the AutoDeletionService will use
the Scheduler to schedule and track files that will be automatically
deleted and identify expired files that are ready for removal.

This is multiple CL change:
  => crrev.com/c/6219361
  -  crrev.com/c/6304078
  -  crrev.com/c/6305579

Bug: 390200553

Change-Id: I59d1650babcc8efa21f06dd9148d1a5f0c53cf9b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6219361
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Daniel White <danieltwhite@google.com>
Cr-Commit-Position: refs/heads/main@{#1426283}