[bookmarks] Fix CHECK failure upon bookmarks cache GUID mismatch
Starting with https://crrev.com/c/6269395, the rare case of detecting a
cache GUID mismatch during startup doesn't work well, and can lead to
CHECK failures. This is because the patch alluded to above overlooked
the codepath that invokes BookmarkModel::RemoveAccountPermanentFolders()
from BookmarkDataTypeProcessor::ConnectIfReady(), which could be
executed too early, if Sync happened to start faster than
BookmarkModel's initialization.

As a simple solution to this, the call to
BookmarkDataTypeProcessor::ConnectIfReady() can be slightly delayed,
unifying the behavior with the more common case where Sync's startup is
slower than BookmarkModel's.

In this patch, this is achieved by posting a task instead of
synchronously invoking BookmarkDataTypeProcessor::ConnectIfReady().

Change-Id: Ia09981c0f5b3636d735b46b8ab122f3c4278d5d9
Bug: 399072142
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6308893
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Rushan Suleymanov <rushans@google.com>
Cr-Commit-Position: refs/heads/main@{#1426287}