Document why AOSP Cronet uses Chromium protoc
I spent a ridiculous amount of time trying to get AOSP protoc to work
for AOSP Cronet, only to be met with linker errors. It then dawned on me
that this can never work due to libc++ ABI incompatibility.

What's even more ridiculous is we as a team already knew this was the
reason why we use Chromium protoc, but apparently we... forgot. Sigh.

Hopefully with this CL, it will be harder to forget.

Bug: b:397933339
Change-Id: I1344d3f40ac23dc76ad1be9ac1ce96671cb03bcd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6309204
Commit-Queue: Etienne Dechamps <edechamps@google.com>
Auto-Submit: Etienne Dechamps <edechamps@google.com>
Reviewed-by: Mohannad Farrag <aymanm@google.com>
Cr-Commit-Position: refs/heads/main@{#1426290}