[iOS][MIM] Cleanup AccountMenuCoordinator
This patch removes this list of methods:
- (void)cancelSignoutAndRecordHistogram:
      (SignoutDataLossAlertReason)reason;
- (void)cancelSignout;
- (void)signoutWithReason:(SignoutDataLossAlertReason)reason;
- (void)signout;

And replaces those methods by one:
- (void)signoutConfirmationWithContinue:(BOOL)continueSignout;

The goal is to be able to extract
`startActionSheetCoordinatorForSignout` into a helper function, and to
be able to share this implementation with AuthenticationFlow.
This will simplify the callback with the helper function.

There is no behaviour changes with this patch, but now the
`Signin_Signout_ConfirmationRequestPresented` user action is now
always recorded when the sign-out confirmation dialog is presented.
This is not an issue for statistic since MIM phase 1 has not been
launched yet.
And this is a good thing to debug crashes.

Bug: 375604649
Change-Id: I5a623e4e29bb1b5f15122533f6c1bdb53f47aec1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6313035
Commit-Queue: Jood Hajeer <jood@google.com>
Auto-Submit: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: Jood Hajeer <jood@google.com>
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1426293}