ozone/drm: always add disable modeset on non-primary tiles for non-tiled modes
Currently, the HardwareDisplayController for a tiled display checks
that:
1. The incoming mode is a non-tiled mode
2. The connector in question is a non-primary connector in the
   HardwareDisplayController
3. The CrtcController is currently enabled
for each connector and only issues a disable if all the above are true.

VT2 currently modesets all connectors for the tiled display to the same
mode. When Chrome re-takes the DRM master, the pipes are left in that
state, but CrtcController does not reflect the enabled state, so no
disable request is issued on non-tiled modes.

This CL changes the HDC to always issue disable request for non-primary
tiles on non-tiled modes.

Bug: b:396696078
Test: Manual testing switching between VT2 on non-tiled mode w/ a tiled display.
Test: ozone_unittests
Change-Id: Id00921c87d57b5377ee79128d20971b2b36da1ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6306187
Reviewed-by: Gil Dekel <gildekel@chromium.org>
Commit-Queue: Su Hong Koo <sukoo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1426341}