[iOS][SigninPromo] Refactor signin fullscreen promo eligibility criteria
This CL essentially moves some eligibility criteria from the
SceneController to the SigninFullscreenPromoSceneAgent and dismisses
others, including duplicate checks.
The following conditions will be moved to the sign-in SceneAgent:
- Checking if the UI is currently blocked
- Checking the sign-in utils conditions (ShouldPresentUserSigninUpgrade)
and the dismissed conditions are the following:
- The current page is the tab grid
- The current page is the settings page
- The app was opened via a URL
For all the other conditions, we either get them for free from the promo
manager or they are duplicate checks from ShouldPresentUserSigninUpgrade.

Bug: 396111368
Change-Id: I76416d572cec1355bad2416dc0839b316675472a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6284666
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Robbie Gibson <rkgibson@google.com>
Commit-Queue: Cheick Cisse <cheickcisse@google.com>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1426344}