Check for invalid RenderFrameHost in CreateHistoryAddPageArgs
In HistoryTabHelper::CreateHistoryAddPageArgs, we obtain the
RenderFrameHost associated with our NavigationHandle. We do this to get
its StorageKey and determine if the navigation is ephemeral.

However, there are some test cases in HistoryTabHelperTest which call
CreateHistoryAddPageArgs with a valid NavigationHandle but an invalid
(mocked) RenderFrameHost, causing the test to crash.

To prevent this issue and any potential in-the-wild crashes, I have
added a simple null-pointer check for the RenderFrameHost we obtain.

Change-Id: Ida35956bf95618b4c7363bbac08e9b6535b7b698
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6309984
Commit-Queue: Kyra Seevers <kyraseevers@chromium.org>
Reviewed-by: manuk hovanesian <manukh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1426383}