[iOS] Implement AutoDeletionService::ScheduleFileForDeletion
In this CL, the AutoDeletionService received additional functionality
that allows it to schedule the user's downloaded content with the
Scheduler for automatic deletion one month from the downloaded date.

In a follow up CL, the AutoDeletionService will add the ability to
remove the downloaded content from user's device once the data is
expired.

This is multiple CL change:
  -  crrev.com/c/6219361
  => crrev.com/c/6304078
  -  crrev.com/c/6305579

Bug: 390200553
Change-Id: I9473cc4962960399bec4fe91152d47b8d7307b61
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6304078
Commit-Queue: Daniel White <danieltwhite@google.com>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1426399}