[Glic] make button square v2
This also removes the margin from the rightmost button in the
tab_stip_action_container via setup button properties. Which was causing
the glic button to be not square when it was that right most button. To
maintain previous behavior the margin was then set during
ExecuteShowTabStripNudge in the cases where it is needed.

Bug: b:398867442
Change-Id: I0677e1755edd8cf829b253e11137449caf7ba963
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6306945
Commit-Queue: Jeffrey Cohen <jeffreycohen@google.com>
Reviewed-by: Emily Shack <emshack@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1426418}