[dtc] Wait for DOM to settle before DTC test reads `chrome://policy`
`device_trust_ui_test.py` clicks "Reload policies" in `chrome://policy`,
then immediately tries to get elements from the DOM. This is inherently
racy because an in-progress reload can replace elements at any time,
which invalidates handles that the WebDriver client holds. There are
up to 20 retries, but they don't cover the refresh of the "Status"
legend [0].

The refresh is very likely to be done after waiting a bit, so borrow
from the client certs test to do that instead.

[0]: https://logs.chromium.org/logs/chrome/buildbucket/cr-buildbucket/8721912019924178193/+/u/find_failed_tests/connector.device_trust_connector.device_trust_connector_windows_enrollment_test.DeviceTrustConnectorWindowsEnrollmentTest/test.py_output

Bug: 327797500, 399483772
Change-Id: I4f3392e7077d6474a2f91fdffeca8646c5d72bb0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6306860
Reviewed-by: Xiang Ji <jxiang@google.com>
Commit-Queue: Jonathan Lee <jonathanjlee@google.com>
Cr-Commit-Position: refs/heads/main@{#1426438}